+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Unreleased]
+### Changed
+* Add note and workaround to README about Poetry error on Ubuntu 22.04.
+  * Pull request [#48](https://github.com/DeepLcom/deepl-python/pull/48)
+    thanks to [FOehlschlaeger](https://github.com/FOehlschlaeger).
+
 ## [1.9.0] - 2022-07-07
 ### Added
 * Add `Translator.create_glossary_from_csv()` allowing glossaries downloaded
@@ -183,6 +189,7 @@ Version increased to avoid conflicts with old packages on PyPI.
 Initial version.
 
 
+[Unreleased]: https://github.com/DeepLcom/deepl-python/compare/v1.9.0...HEAD
 [1.9.0]: https://github.com/DeepLcom/deepl-python/compare/v1.8.0...v1.9.0
 [1.8.0]: https://github.com/DeepLcom/deepl-python/compare/v1.7.0...v1.8.0
 [1.7.0]: https://github.com/DeepLcom/deepl-python/compare/v1.6.0...v1.7.0